requirements.txt
Django
requests-oauthlib
python-openid
requests
pyjwt
python3-saml
djangorestframework
Pillow
django-allauth
celery
redis
dj-database-url
psycopg2-binary
boto3
django-storages
gunicorn
whitenoise
django_celery_results
django_extensions
"django-allauth[socialaccount]"